requirements.txt
autogen-agentchat
autogen-ext[ollama]
python-dotenv